Move function to where it is used
